Checking in original data model for Assembly Theory.
This seems to work OK, now I need an algorithm that works for general
objects.

The problem with only representing objects as strings is that the
assembly theory space isn't linear, it's a tree. The structure needs
to be nested in order to represent the full complexity of the concept.