Modifying innards of the Assembly object to use a graph instead of str.
This relies on the networkx library to do. As a first commit, I assume a
linear, ordered, object like a string but this will be extended to
support other graphs, too. This is quite slow compared to before, but
more versitile. I'm hoping to leverage networkx's vast library of
algorithms to help the optimization process.

First idea: instead of bottom-up path construction, do a top-down path
construction instead. I think this will be ~as fast but there may be
cleverer path pruning optimizations. There's a 200x speed decrease in
this change that I gotta fix.